Added comments + fixed upload regex
Added some comments in the file upload POST request code in api/file/route.js and the function which makes the POST request in page.js. Also fixed the regex in the route.js file (added the '$' at the end) so it only matches file extensions at the end of the word (just in case users upload files containing periods in the name before the extension).